feat: Refactor kdtree_analytics.go and create API audit
Decomposed the "God Class" `kdtree_analytics.go` into three distinct files:
- `kdtree_analytics.go`: Core tree analytics
- `peer_trust.go`: Peer trust scoring logic
- `nat_metrics.go`: NAT-related metrics

Renamed `ComputeDistanceDistribution` to `ComputeAxisDistributions` for clarity.

Created `AUDIT-API.md` to document the findings and changes.
